+++ test/benchmarks/raphael/complex-svg/data.js
