+++ test/benchmarks/raphael/complex-svg/fabric.js
