+++ test/benchmarks/raphael/complex-svg/raphael.js
