+++ test/benchmarks/raphael/complex-svg/testem.js
@@ -1,6 +1,7 @@
-const config = require('../testem.raphael');
+const config = require('../testem');
+const path = require('path');
 
-const entry = 'test/benchmarks/raphael/tiger';
+const entry = path.relative(process.cwd(), __dirname);
 
 module.exports = {
   ...config,
